requirements.txt
psycopg2-binary
scrapetube
youtube-transcript-api
openai
pinecone
python-dotenv
tqdm
fastapi
uvicorn
anthropic
yt-dlp
striprtf
